fix: add eventlistener when update chat page
