+++ src/Public/global.js
@@ -1,4 +1,4 @@
-export const FRONTEND = "http://127.0.0.1:5001";
+export const FRONTEND = "http://127.0.0.1:5500";
 export const BACKEND = "http://127.0.0.1:8000";
 export const LOCALHOST = "127.0.0.1";
